+++ src/main/java/minefantasy/mf2/block/crafting/BlockAnvilMF.java
@@ -69,7 +69,7 @@ public class BlockAnvilMF extends BlockContainer
     @Override
     public void onBlockPlacedBy(World world, int x, int y, int z, EntityLivingBase user, ItemStack item)
     {
-    	int dir = MathHelper.floor_double((double) (user.rotationYaw * 4.0F / 360.0F) + 0.5D) & 3;
+    	int dir = MathHelper.floor_double(user.rotationYaw * 4.0F / 360.0F + 0.5D) & 3;
 
         world.setBlockMetadataWithNotify(x, y, z, dir, 2);
     }
@@ -142,7 +142,7 @@ public class BlockAnvilMF extends BlockContainer
                         }
 
                         itemstack.stackSize -= j1;
-                        EntityItem entityitem = new EntityItem(world, (double)((float)x + f), (double)((float)y + f1), (double)((float)z + f2), new ItemStack(itemstack.getItem(), j1, itemstack.getItemDamage()));
+                        EntityItem entityitem = new EntityItem(world, x + f, y + f1, z + f2, new ItemStack(itemstack.getItem(), j1, itemstack.getItemDamage()));
 
                         if (itemstack.hasTagCompound())
                         {
@@ -150,9 +150,9 @@ public class BlockAnvilMF extends BlockContainer
                         }
 
                         float f3 = 0.05F;
-                        entityitem.motionX = (double)((float)this.rand.nextGaussian() * f3);
-                        entityitem.motionY = (double)((float)this.rand.nextGaussian() * f3 + 0.2F);
-                        entityitem.motionZ = (double)((float)this.rand.nextGaussian() * f3);
+                        entityitem.motionX = (float)this.rand.nextGaussian() * f3;
+                        entityitem.motionY = (float)this.rand.nextGaussian() * f3 + 0.2F;
+                        entityitem.motionZ = (float)this.rand.nextGaussian() * f3;
                         world.spawnEntityInWorld(entityitem);
                     }
                 }
