+++ src/main/java/minefantasy/mf2/item/weapon/ItemBattleaxeMF.java
@@ -63,7 +63,7 @@ public class ItemBattleaxeMF extends ItemHeavyWeapon
 		
 		if(lunge)
 		{
-			float fallBonus = ((float)Math.min(entityHitting.fallDistance, baseDamage))/2F;
+			float fallBonus = (Math.min(entityHitting.fallDistance, baseDamage))/2F;
 			if(entityHit.onGround && !properHit)
 			{
 				entityHitting.hurtResistantTime = 20;
